EH-747: Moved AiemminHankitunYTOOsaAlue.ts to a new folder
