+++ shared/models/YhteinenTutkinnonOsa/AiemminHankitunYTOOsaAlue.ts
@@ -1,12 +1,12 @@
 import { types } from "mobx-state-tree"
-import { OsaamisenOsoittaminen } from "./OsaamisenOsoittaminen"
-import { TodennettuArviointiLisatiedot } from "./TodennettuArviointiLisatiedot"
-import { KoodistoVastaus } from "./KoodistoVastaus"
-import { AiemminHankitutTutkinnonOsatViews } from "./helpers/AiemminHankitutTutkinnonOsatViews"
-import { Organisaatio } from "./Organisaatio"
-import { TutkinnonOsaType } from "./helpers/ShareTypes"
-import { EnrichKoodistoKoodiUri } from "./Enrichment/EnrichKoodistoKoodiUri"
-import { EnrichOrganisaatioOid } from "./Enrichment/EnrichOrganisaatioOid"
+import { OsaamisenOsoittaminen } from "../OsaamisenOsoittaminen"
+import { TodennettuArviointiLisatiedot } from "../TodennettuArviointiLisatiedot"
+import { KoodistoVastaus } from "../KoodistoVastaus"
+import { AiemminHankitutTutkinnonOsatViews } from "../helpers/AiemminHankitutTutkinnonOsatViews"
+import { Organisaatio } from "../Organisaatio"
+import { TutkinnonOsaType } from "../helpers/ShareTypes"
+import { EnrichKoodistoKoodiUri } from "../Enrichment/EnrichKoodistoKoodiUri"
+import { EnrichOrganisaatioOid } from "../Enrichment/EnrichOrganisaatioOid"
 
 export const Model = types.model("AiemminHankitunYTOOsaAlue", {
   id: types.optional(types.number, 0),
