+++ shared/models/helpers/TutkinnonOsa.ts
@@ -3,7 +3,7 @@ import { OsaamisenHankkimistapa } from "../OsaamisenHankkimistapa"
 import { OsaamisenOsoittaminen } from "../OsaamisenOsoittaminen"
 import { AiemminHankittuAmmatillinenTutkinnonOsa } from "models/AiemminHankittuAmmatillinenTutkinnonOsa"
 import { AiemminHankittuPaikallinenTutkinnonOsa } from "models/AiemminHankittuPaikallinenTutkinnonOsa"
-import { AiemminHankitunYTOOsaAlue } from "models/AiemminHankitunYTOOsaAlue"
+import { AiemminHankitunYTOOsaAlue } from "models/YhteinenTutkinnonOsa/AiemminHankitunYTOOsaAlue"
 import { HankittavaAmmatillinenTutkinnonOsa } from "models/HankittavaAmmatillinenTutkinnonOsa"
 import { HankittavaPaikallinenTutkinnonOsa } from "models/HankittavaPaikallinenTutkinnonOsa"
 import { YhteisenTutkinnonOsanOsaAlue } from "models/YhteinenTutkinnonOsa/YhteisenTutkinnonOsanOsaAlue"
